:racehorse: First cut of initialization benchmarks
Adding AdventureWorks model. Currently the tests that require the database are skipped if you don't have AdventureWorks installed because the artifacts to create it are large (backup is 45mb, SQL script to create schema and insert data is >500mb). We could look at dynamically downloading artifacts and creating it in the future if we think it adds significant value.