Make Add/Attach/Update methods add dependents by default
This gives them a limited form of graph behavior, as discussed in the design meeting. Passing includeDependents: false reverts to the single object behavior.

Note that the params sugar methods for AddRange, etc. do not have the flag (because there is no good place to put it with this signature) and always do the default graph behavior. If the single object behavior is needed, then it can be obtained by using the IEnumerable overload instead of the params overload.

Also, we need to decide whether calling Add/Attach/Update with graph behavior on an already tracked entity causes the state of that entity to change, and, if so, whether the state of dependents should also change. Currently if the entity is already tracked, then its state is changed but nothing else happens.

The behavior of DetectChanges has not yet been updated.