:green_heart: Fix SaveChanges init benchmark
Test would previously insert and entity type that used IDENTITY in the
database. Even though we roll the transaction back the value still gets
reserved in the database and so we were running out of IDENTITY values
after several test runs (the column is `smallint`, so this happened
after a few days). Swapping to use an entity with a non-generated key.