Flow cascade delete from metadata into Migrations
Also, fix R# block indenting back to match what Visual Studio does so that we don't have so many conflicts between R# formatted code and VS formatted code.

Currently the functional test for this is disabled in SQLite since cascade delete for SQLite does not appear to be working at this time.