Fix to include #3163 and #3169
#3163 - Query :: include doesn't work well with inheritance

Problem was that during binding we were assuming that memberExpression.Expression would always be QuerySourceReferenceExpression. However in some cases (inheritance related) it is wrapped in a convert.
Fix is to strip convert from the expression before checking if it's a QuerySourceReferenceExpression

#3169 - Query :: include with inheritance can return fewer results than expected when including optional (null valued) navigations

Problem was that for include we produce LEFT JOINs, however the discriminator predicate that is applied to entities with inheritance (to restrict them to only sub-types requested) was applied to the entire query.
This could cause some rows being filtered out. Fix is to detect where the discriminator predicate is needed and modify the expression accordingly - we create a subquery and stick discriminator predicate on it,
so it only applies to entities that it should.

Also added lot of tests for the area (include + inheritance) to cover other similar cases.

CR: Andrew