Remove persist flag from in-memory database
In API review we decided to allow named in-memory databases. I will file a bug for that and we can do it if we feel it is high enough priority. The easy change for now is to just remove the persist flag and make it such that in-memory databases always persist per model, per service provider, which is what we were doing in almost all our tests anyway.