Fix to #3091 - Include() Appended to Query is Dependent on Order of Joins
Problem was that include annotation always assumed that it's query source is the MainFromClause. In the scenario from the bug it is not the case.
Fix is to fix-up the query source of the annotation, by looking at the referenced query source of the NavigationPath,
e.g. in nav path [c].Orders, c is a QuerySourceReference expression to appropriate query source (the one being included, even if it's not the MainFromClause)
CR: Andrew