Allow setting cascade action of SetNull
This means that a principal with optional dependents can be deleted and without loading the dependents and the FKs will be set to null automatically by the database.

This is not turned on by default because SQL Server will frequently error out indicating that it generates cycles in the schema.