:bug: Fix ServiceProviderCache Hash
The hashing algorithm for `ServiceProvider` was using the hash code on
the instance for anything that was added as an instance. This was broken
when we started adding `TelemetryListener` because we were adding a
different instance each time `AddRelational()` was called. This showed
up as a massive slow down in our perf tests because we effectively
stopped caching anything when the context is responsible for managing
the `ServiceProvider`.

Two changes as part of this:
* Stop using the actual instance when calculating the hash and just use
the type.
* Swap to using `AddSingleton` instead of `AddInstance` for
`TelemetryListener`. Not strictly needed due to the previous change, but
it seems cleaner.