:racehorse: Widen tables for benchmarks
The domain model used for many of the benchmarks only had a few columns
per table, which is probably not representative of the most common
scenarios that EF is used in. This expands them to have more columns.

I did look at using AdventureWorks for all the tests (we use it for the
initialization tests), but it gets hard to zero in on a particular
scenario using the data patterns and counts that are present in the
database.