Fix to #3103 - Must be reducible node exception when performing Join
Problem:
We were incorrectly expanding navigations that appeared in the inner key selector of a join expression (right side).
Normally we expand nav props into additional join. E.g.:

from p in ctx.Payouts
join a in ctx.Assets on p.Good.Id equals a.Id
select p

converts to:

from p in ctx.Payouts
join g from ctx.Goods on p.Id equals g.Id
join a in ctx.Assets on g.Id equals a.Id
select p

However if navigation propery is in the inner key selector, like so:

from p in ctx.Payouts
join a in ctx.Assets on p.Id equals a.Good.Id
select p

we encounter chicken-and-egg problem:

from p in ctx.Payouts
join g in ctx.Good on a.Id equals g.Id // invalid, because a was not declared yet
join a in ctx.Assets on g.Id equals a.Id

or

from p in ctx.Payouts
join a in ctx.Assets on g.Id equals a.Id // invalid, because g was not declared yet
join g in ctx.Good on a.Id equals g.Id

Solution:

Some navigation properties can be translated into FKs, which removes the problem:

from c in Customers
join d in Details on c.Id equals d.Customer.Id
select c

can be translated into:

from c in Customers
join d in Details on c.Id equals c.CustomerId
select c

This can only be done for a subset of queries - join needs to be based on key property, FK must be pointing the right way, and key must not be composite.

For all other cases we translate this into a subquery:

from c in Customers
join d in Details on c.Id equals
(from c2 in Customers
 where c2.CustomerId equals d.Id
 select c2.Id).FirstOrDefault()

In case of nested navigation property we expand the first one into a subquery, and add extra joins to the subquery, just like we did originally. Only the first navigation is "problematic", the following ones can be safely translated to joins
CR: Andrew