Replace ThreadSafeDictionaryCache with ConcurrentDictionary
ThreadSafeDictionaryCache was implemented on top of ImmutableDictionary, possibly before ConcurrentDictionary was available, but also for perceived performance gains especially where reading from the cache (as opposed to adding to it) dominates. However, perf testing is showing the ConcurrentDictionary is considerably (30-40%) faster even on read-heavy scenarios, so replacing to use this instead.