Move caching of property access delegates to the model
This is because the dictionary access involved in looking up the delegates is a significant perf hit, and co-location of the accessors with the property/navigation for which they are used removes this overhead.

Since he model is not coupled to the service provider, this means that the delegate construction code is now not a replaceable service. The mitigation for this in the long term is to allow extensibility of the model to contain delegates created in different ways. For example, the factories used to create the delegates could be annotated onto the model.

Note that the delegates are not stored as annotations because the lazy creation needs to happen after the model is "built" and must be thread-safe, neither of which are supported by annotations. Also, this is pert-sensitive code and the lookup needs to be fast.