Fix to Issue #3452 - Error when return complex linq to json object The EF.Property<T> method may only be used within LINQ queries.
Problem was that we incorrectly that relationship would flow in one direction always assigning foreign key property to the outer key selector and primary key to the inner selector for the comparison.
Fix is to check the direction of the navigation and correctly construct the MemberExpressions. Also making sure we compensate for differences in nullabilities between keys, when necessary.