Avoid checking fixup in state manager when query is known to have done it
Query always fixes up navigations between entities returned by a given query, regardless of whether the query is going to be tracked or not. If the state manager ends up tracking only these entities, then it does not need to do any fixup work. Not doing this work can improve query tracking performance by about 22% for this common case.

The state manager switches out of this mode when:
- A second query starts
- An entity is attached or added
- An entity changes to the Modified or Added state, since this could be an FK change