Fix to #3499 - Exception with in-list where clause across joined table
Problem was that during navigation expansion we would add the created join clauses into the query model that the navigation was used. However, sometimes the navigation references a query source from the outside query, and the joins should be added there instead.

Fix is to keep track of hierarchy of query models and testing whether the given query model is the right one to add the join to (i.e. it contains a declaration of the query source that the join references)

Also added some minor fixes - we were not performing navigation expansions for result operators and we were not translating EF.Property() inside Contains result operator properly.
CR: Andrew