Fix to #3758 - Including Many-To-Many Relation in Select
Problem was that we were making some incorrect assumptions when materializing collections directly - we always assumed there will be a subquery present in those cases, and also we were not materializing correct collection types.
Fix is to inject subquery to navigation access, if the subquery was not present before. Also added a client-side method that materializes a correct collection, based on the information in the INavigation.
CR: Andrew, Arthur