Fix a bug with concatenation of string and non-string, (e.g. c.Name + 5)
Problem is twofold - compiler translates this into a BinaryAdd operation, with a string Concat(object, object) method info, and casts the int into object. When we translate query and encounter a object-valued constant we don't know how to translate it. Another issue is that when generating SQL, we need to explicitly cast the non-string to varchar (so that types match). Currently we are always ignoring casts (were treating them as means to have consistently typed expression trees, but they were not needed in actual SQL generation).

Fix is to add fragment translator for the BinaryExpression.Add that recognizes concatenation of string with non-string, unboxing the non-string constant and wrapping it in a (new) ExplicitCast node. This node is then translated into proper CAST ([non-string] AS [string-type]). String-type name is determined using appropriate RelationalTypeMapper

CR: Diego