Remove GraphBehavior enum and track all reachable entities
Issue #4424

This makes the behavior for Add/Attach/Update be to track all reachable entities, as in EF6. Single object operations can be done with the EntityEntry API. In the future we made add a context-wide flag to modify this behavior.

Also, make Remove still be single object, but do attachment of reachable entities such that cascade deletes will be processed correctly on SaveChanges.