Remove GraphBehavior enum
This leaves the delete children (dependents) behavior as the default. Single object operations can be done with the EntityEntry API. In the future we made add a context-wide flag to allow dependents and principals to be added, or to allow single objects to be added with this API.

Also, make Remove still be single object, but do attachment of children such that cascade deletes will be processed correctly on SaveChanges.