Ensure notification entities brought in by query are tracked as Unchanged
Issue #4020

The issue is that when query sets the navigation property this is detected by the change detector and the new entity is marked as Added. Fix is for query to disable change detection when it is setting navigation properties. This creates a coupling between query and change detection which is unfortunate, but perhaps cleaner than letting the change detector cause a state change and then correcting it later.