:memo: Improve HasKey on derived type exception message
Resolve #3863

A common cause of this exception is unintentionally including a base
type in the model. Adding info about how to stop this happening to the
exception message.

**Old Message:** The derived type 'Address' cannot have keys other than
those declared on the root type 'AddressBase'.

**New Message:** A key cannot be configured on 'Address' because it is a
derived type. The key must be configured on the root type 'AddressBase'.
If you did not intend for 'AddressBase' to be included in the model,
ensure that it is not included in a DbSet property on your context,
referenced in a configuration call to ModelBuilder, or referenced from a
navigation property on a type that is included in the model.