Use type mapper correctly in SQL Server batch updates
See Issue #4399

The issue was that the update pipeline was always asking the type mapper for the type to use even if a type had been explicitly set for the property. Fixed this and added tests for inserting all mapped types as both single inserts and multiple inserts both with and without Identity columns.