Use DbSet property name as table names
Issue #4540

The DbSet property name from the context is used as the name of the table mapped to by that entity type.

If the table name is specified explicitly with ToTable, then this is used instead.

The entity type name is used as the table name for entity types in the model that do not have a corresponding DbSet property. If there is more than one DbSet property for the same entity type, then we fall back to using the entity type name.