Update to Reduce magic in the interaction of EF Core and DI
See issue #4668. This is an update following a discussion yesterday that moves all configuration onto DbContextOptions and makes this the only object passed to the DbContext constructor. This allows the builder pattern that already exists for DbContextOptions to be used for external services and for setting the internal service provider. AddDbContext does this automatically for external services, but a DbContextOptions object can also be easily built and either configured in DI or passed directly to the context with new.

In addition, the UseMyProvider methods have been updated to use nested closure for provider-specific configuration. This makes for a very nice chaining story.

There is a new overload of AddDbContext that makes the application service provider available to teh configuration delegate. This is rarely needed, but one use is to set the application service provider to be used as the internal EF service provider if this is needed/desired.

Still to do:
- Look at negative cases when services are not setup and create better exception messages.
- Rename AddSqlServer, etc to AddEntityFrameworkSqlServer, etc.
- Extract diagnostics as one of the external services on DbContextOptions.