Throw helpful exceptions when services or provider are not configured
The common case here is a context that does not have OnConfiguring and does not pass in a DbContextOptions to its constructor, or the DbContextOptions passed does has not had any providers configured on it.

It is also possible to get no EF services at all, but this now requires making a call to UseInternalServiceProvider with a service provider that does not have EF services in it.

Also fixes #4665.