Make Attach never have side effect of setting dependent entities as Modified
This means that graph Attach calls behave as expected no matter which entity is attached first or which mechanism is used, and work correctly with shadow FKs.

It also means that re-parenting an attached dependent cannot be done by setting a new principal to point to it, and then attaching that principal. But attaching the principal and then setting the navigation property will work.