Introduce an Undefined state for property values
Shadow properties get this state when entities are first attached. These properties are then not considered modified even if the entire entity is marked as modified and so the values for these properties are not saved. The properties are only considered as modified or not when an actual value is set.

See #4474