Split storage of default value SQL and computed value SQL
Because not all provider support both concepts we were storing these things as a single string and then letting other metadata determine how to use the value. However, for cases like that in issue #4501 Migrations assumes a computed column because it doesn't know that the value specified is actually a default and not computed expression. The fix is to stop combining the metadata.