Ensure entities from query are tracked as correct type
Issue #4817

The problem was that query was passing in the base type to the state manager since the actual type varies by row. It would likely be possible to flow the actual type through from the materializer but this would require significant changes and would not be used in a lot of cases. Instead the state manager now looks up the entity type if it is not the base type, with some optimizations to that lookup made in metadata.