Throw when using multiple context types with non-generic DbContextOptions
Probably root cause if issue #4994.

If there are multiple context types registered in D.I. and each depends on non-generic DbContextOptions, then only one of them will get the correct instance. This fix checks that a context is getting its correct options and throws if not with the guidance to use generic DbContextOptions parameters in the context constructors.