Handle null/empty in property name in notification events
See #3625. Passing in null or the empty string indicates that all properties are changing/changed. We now honor this, except for properties marked as read-only. We ignore these since we don't track original values for full notification entities and hence have no way of knowing if they are really changed or not, and we cannot mark them as modified. So we assume that since they are read only they have not been changed.