Prevent discriminator columns from failing with in-memory database
Issue #4285

The default, provider-agnostic, value generator was setup to always generate temporary values. This was a carry over from when we required this so that entities could be added in any order, but that behavior was changed and so always generating temporary values is no longer needed.

The default value generator now doesn't generate temporary values, which may not be handled by the provider--such as is the case for in-memory. Where possible default value generators are still supplied that make use of GUIDs. One of these now gets used in this scenario.

This change doesn't really fix the problem that structural changes made to the model by a provider may not be handled correctly by other providers, but it fixes the issue at hand.