Throw when attempting to access original value when it is not being tracked
Issue #4375.

For entities that implement INotifyPropertyChanging, the original value is not needed by changing tracking and hence we do not store it or allocate space to store it. Previously if you asked for the original value in such cases we would give you back the current value. We now throw saying that the value is not available and directing you to the flag that will make it available. It's a pretty blunt flag right now, but we can make it more fine-grained if enough people want this behavior.

In fixing this, it turns out that the update pipeline was using the original value for primary key where clauses. This started to throw because the original value was not available. This has been changed to use the current value instead. This was complicated because the logic for choosing original or current was implemented twice and both needed to match, so this has been changed to use a common implementation.