Make notification change tracking opt-in and implement collection notifications
Issue #5075
Issue #445
Also, updates options for PR #5118

Use of INotifyProperyChanged and INotifyPropertyChanging is now opt-in. If you opt-in, then you must also start using INotifyCollectionChanged on your collection nav props, which is now implemented. We check these things in various places, including model validation and when we start tracking entities to try to make sure that you won't think that change tracking is happening when it isn't.

Doing a Clear on ObservableCollection does not work because the event does not tell us which items were removed. This has always been a problem for EF. An implementation of ObservableCollection where Clear does work is provided, and is used by default when we initialize collections on notification entities.