Implementation of ObservableHashSet
This is an implementation of INotifyCollectionChanged that is backed by a HashSet instead of a List. This will then perform better in Contains when there are a large number of elements, which is something that ObservableCollection doesn't do well with. It also implements Clear in a way that can be used by EF notifications because the items being cleared are passed in the event.

Copies the API surface of HashSet<T>.

Also implements INotifyPropertyChanged/Changing for the Count property.

Does not have an inherent ordering and so does not have support for indexers, etc., or the notifications that go with these.