Fix to #5043 - Query: Navigations: Using EF.Property with reference nav omits null compensation
Problem was that navigation expansion logic was only applied for MemberExpressions, but never for MethodCallExpression. Fix is to abstract that logic to work for both MemberExpressions and MethodCallExpression that happen to be EF.Property.
This enables using regular member access and EF.Property interchangeably in the context of navigation properties.

Also did some minor bugfixes and optimizations found along the way:
- compensating for nullability differences in SqlTranslatingExpressionVisitor, when processing comparison,
- preventing EntityEqualityRewritingVisitor to short-circuit on non-equality operations in case of nested binary expressions,
- removed some redundancy in Member/Method binding code,
- making Single and SingleOrDefault always evaluate on the client, to better mimic L2O behavior.

CR: Andrew