Fix to #3101 - .Include() with Self-Join / Coalesce
Problem was that for include we were using relinq's EntityAccessorFindingVisitor, to extract a accessor lambda to a given query source in the selector (for InMemory scenario). Problem was that EntityAccessorFindingVisitor was not meant to work for some cases that we were using it (e.g. Coalesce, Conditional).

Fix is to refactor InMemory include mechanisms so that we no longer need the accessor lambda. Now Include method is applied either around appropriate elements of the outer-most selector, or in case no Select method has been generated - around EntityQuery/OfType directly.

CR: Andrew