:memo: Don't assume failed GetService means no provider
If you call an extension method such as
`DbContext.Database.GetDbConnection()` and we fail to resolve the
connection from DI, then we assumed this was because you did not
register a provider. This is often the case, but it's a misleading error
message if you are just trying to resolve some service that is not
present. Adding some pre-amble to this exception to explain this.

Note we still throw the existing message if we are trying to locate the
configured provider and there is none
(see `DatabaseProviderSelector.SelectServices()`) since we are sure that
is the problem in that case.

Fixes #5500