Give up sooner if Exists is experiencing long timeouts
Issue #4454. After we create a database we wait for it to report as created because of the notorious issues with SQL Client/SQL Server and no process at end of pipe, etc. However, if the timeout takes a very long time, then the retry logic could make this a very, very long time. So now we give up if it seems like we have waited more than a minute and we are still getting failures.