Fix to #5571 - RC2 Upgrade Eager Loading throw an exception
Problem was in the way that we were doing PushDownSubquery during complex includes. We were incorrectly overriding aliases of the projection elements from the outer query in some cases. Basically, we were looking at the aliased column name to determine whether it needs to be aliased again. Instead we should look at the alias first, and only if that is not set or it would be a duplicate, we should re-alias the column.

CR: Smit