Fix to #4739 - UseRowNumberPaging causes NRE in complex queries
Problem was that during query rewrite for RowNumberPaging, when copying projection elements from the inner query, we were not preserving aliases. Instead were using ColumnExpression (so it would be the projected element in the outer query) even in the alias was explicitly set in the inner SelectExpression. Fix is to try use alias first and only if alias is not set fall-back to using the column name.

CR: Smit