Fix to #5630 - Select Distinct with where/Any subquery always includes key from joined table, and returns too many rows
Problem was that for correlated subquery when we visit the element that references the outer query, we add that element to the root projection. We need that potentially if the query ends up being client-evaluated. However, if client-eval is not needed, those additional projection elements can be thrown away. Currently they were not, resulting in unnecessary columns being pulled from the database, or in some cases (e.g. Distinct) could lead to incorrect results being returned.

Fix is to do a snapshot of projections before we start visiting the subquery and if it ends up being fully translated to SQL, we prune those additional projection elements.

CR: Andrew