Sugar for HasForeignKey/HasPrincipalKey
Issue #5377

Added back generic overloads so that in the common case of a shadow property in CLR entity types the generic method can be used instead of using typeof().

Also, only for these APIs, the short name of the type can be passed as a string since the resolution space is very small--it just checks the name at either end of the relationship.